Checkpoint: test refactor for auth.test.ts
