+++ src/index.ts
@@ -2,6 +2,6 @@ export * from './types.js';
 export * from './payMcpClient.js';
 export * from './solanaPaymentMaker.js';
 export * from './oauthClient.js';
+export * from './oauthResourceServerClient.js';
 export * from './oauthClientDb.js';
 export * from './auth.js';
-export * from './authOAuthHelpers.js';
